requirements.txt
torch==2.7.0
torchvision==0.22.0
pytorch-lightning==2.5.1.post0
scikit-learn==1.6.1
matplotlib==3.10.3
seaborn==0.13.2
pandas==2.2.3
kagglehub==0.3.12
Pillow==11.2.1
numpy==1.26.4
scipy==1.15.3
tensorboard==2.19.0
tensorboardX==2.6.2.2
typing_extensions==4.13.2
xgboost==3.0.2
onnx==1.18.0
onnxruntime-gpu==1.22.0
tritonclient[http]
